hub: extract WebSocket message dispatch into ws/{client,runner}Message (#51 phase 2)
Phase 2 of the #51 hub.ts split. Pure structural refactor on top of
phase 1 (#53). All 89 hub tests pass without modification.

Extractions:

hub/src/ws/context.ts
  WsContext interface: db, metrics, pendingHistory, runners map, plus
  bound sendToClient/sendToRunner/relayToClients/broadcast helpers and
  enrichedMachineList. Built once per dispatch by Hub.wsContext() so
  the literal closes over `this` and Hub internals stay private.

hub/src/ws/clientMessage.ts
  handleClientMessage switch + the per-message dispatchers it routes
  to (handleStartSession, handleSendPrompt, handleRespondToPrompt,
  handleListSessions, handleListMachines, handleGetSessionHistory)
  + resolveSessionRunner helper + isSafeDirectory util (it's only used
  here, so it lives next to its caller now).

hub/src/ws/runnerMessage.ts
  handleRunnerMessage switch (stream/tool/user_prompt relays,
  session_history reply with peek/take + degraded counter, session_*
  status updates, dropped_tool_block counter).

Connection lifecycle (handleClientConnection, handleRunnerConnection)
stays in hub.ts because it's tightly coupled to the connection setup
(auth, addClient/removeClient, ws.on bindings) and is short.

Result: hub.ts dropped from ~700 to ~480 lines. Combined with phase 1
that's a ~45% reduction from the original ~860-line god-object. The
remaining hub.ts is the wiring layer: constructor, lifecycle, HTTP
routing, WebSocket connection setup, and the per-Hub helper methods
(sendToClient/sendToRunner/relay/broadcast/enrichedMachineList) that
the WS context binds to.

Closes #51.
